Added prefabs

+++ BuildingGenerator/Shared/Transform.cs
@@ -19,8 +19,8 @@ namespace BuildingGenerator.Shared
 
 
         // TODO: ezeket valahogy kalkulált mezőkké tenni, nem megadottnak
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public float Width { get; set; }
+        public float Height { get; set; }
 
         /*
         Vector3[] Normals { get; set; }
